+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.1.4 (2015-11-12)
 ----------------
 
--  Enhance Create Ticket functionality
+-  Add `create` class method to the Ticket model; Add `Contact` model.
 
 0.1.3 (2015-09-13)
 ----------------